Full Stack Budgeting Software Application/backend/requirements.txt
asgiref==3.8.1
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.0.2
django-allauth==65.3.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.4
PyJWT==2.8.0
oauthlib==3.2.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.12.2
urllib3==2.0.4
python-dateutil==2.8.2
